requirements.txt
streamlit
opencv-python
numpy
pandas
matplotlib
seaborn
Pillow
plotly
